requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
torch
torchvision
pydantic
streamlit
pillow
requests